NEW: jpa 상속버전 레파지토리 추가
